structure: improve code structure
