+++ Cargo.toml
@@ -12,3 +12,9 @@ serde_json = "1.0.68"
 tokio = { version = "1.0", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+
+[[bin]]
+name = "web"
+path = "./src/bin/web/main.rs"
+
