+++ src/domain/commands.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct CreateUser {
+    pub username: String,
+}
