+++ src/domain/models.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct User {
+    pub id: u64,
+    pub username: String,
+}
